remove submodules, one top level module only

+++ mir_dataset_loaders/download_utils.py
@@ -7,10 +7,22 @@ import zipfile
 
 from tqdm import tqdm
 
-from .. import MIR_DATASETS_DIR
+from . import MIR_DATASETS_DIR
 
 RemoteFileMetadata = namedtuple('RemoteFileMetadata',
                                 ['filename', 'url', 'checksum'])
+
+
+def get_save_path(data_home):
+    if data_home is None:
+        save_path = MIR_DATASETS_DIR
+    else:
+        save_path = data_home
+
+    if not os.path.exists(save_path):
+        os.makedirs(save_path)
+
+    return save_path
 
 
 def md5(fname):
@@ -28,7 +40,7 @@ class DownloadProgressBar(tqdm):
         self.update(b * bsize - self.n)
 
 
-def download_from_remote(remote, data_home=None):
+def download_from_remote(remote, data_home=None, clobber=False):
     """Download a remote dataset into path
     Fetch a dataset pointed by remote's url, save into path using remote's
     filename and ensure its integrity based on the MD5 Checksum of the
@@ -49,18 +61,19 @@ def download_from_remote(remote, data_home=None):
     file_path: string
         Full path of the created file.
     """
-    # TODO: Check if download_path already exists?
     download_path = (
         os.path.join(MIR_DATASETS_DIR, remote.filename) if data_home is None
         else os.path.join(data_home, remote.filename)
     )
 
-    with DownloadProgressBar(unit='B', unit_scale=True,
-                             miniters=1, desc=remote.url.split('/')[-1]) as t:
-        urllib.request.urlretrieve(
-            remote.url, filename=download_path, reporthook=t.update_to)
+    if not os.path.exists(download_path) or clobber:
+        # If file doesn't exist or we want to overwrite, download it
+        with DownloadProgressBar(unit='B', unit_scale=True,
+                                 miniters=1,
+                                 desc=remote.url.split('/')[-1]) as t:
+            urllib.request.urlretrieve(
+                remote.url, filename=download_path, reporthook=t.update_to)
 
-    # urllib.request.urlretrieve(remote.url, download_path)
     checksum = md5(download_path)
     if remote.checksum != checksum:
         raise IOError("{} has an MD5 checksum ({}) "
@@ -70,7 +83,7 @@ def download_from_remote(remote, data_home=None):
     return download_path
 
 
-def unzip(zip_path, save_dir, cleanup=True):
+def unzip(zip_path, save_dir, cleanup=False):
     zfile = zipfile.ZipFile(zip_path, 'r')
     zfile.extractall(save_dir)
     zfile.close()
@@ -78,7 +91,7 @@ def unzip(zip_path, save_dir, cleanup=True):
         os.remove(zip_path)
 
 
-def untar(tar_path, save_dir, cleanup=True):
+def untar(tar_path, save_dir, cleanup=False):
     tfile = tarfile.TarFile(tar_path, 'r')
     tfile.extractall(save_dir)
     tfile.close()
